% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sheaf.R
\name{create_sheaf}
\alias{create_sheaf}
\title{Create Sheaf}
\usage{
create_sheaf(restriction_maps, source, destination)
}
\arguments{
\item{restriction_maps}{A list of functions. For each pair of open sets there is a restriction.
These can be projections, linear maps, etc.}

\item{source}{A list of originating nodes.}

\item{destination}{A list of terminal nodes.}
}
\value{
The output of this function is a data frame.
}
\description{
create_sheaf is a basic set-up function in sheafr, and the output is a standardized sheaf.
A sheaf is a topological data structure, encapsulated here as a formatted data frame.
This creates a model for data to flow along a path through provided restriction maps, or functions.
}
\details{
create_sheaf
}
\examples{
X <- function(stalk){select("y")}
create_sheaf(restriction_maps = c(X), source = c("A"), destination = c("B"))
}
