% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_assignment.R
\name{read_assignment}
\alias{read_assignment}
\title{Read Assignment}
\usage{
read_assignment(source, values, variables)
}
\arguments{
\item{source}{A list of characters or factors. Each source is an originating node.}

\item{values}{A list of doubles. The values attributed to variables at each node.}

\item{variables}{A list of characters. The variables at each node.}
}
\value{
returns a formatted data frame
}
\description{
This function formats the necessary columns for a sheaf assignment.
This can be paired with sheaves later in your sheaf building process.
}
\details{
read_assignment
}
\examples{
read_assignment(source = c("A", "B", "C"),
values = c(2, 4, 3),
variables = c("x", "y", "z"))
}
