% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheaf_assignment.R
\name{sheaf_assignment}
\alias{sheaf_assignment}
\title{Initiating a Sheaf-Assignment Pair}
\usage{
sheaf_assignment(assignment, sheaf)
}
\arguments{
\item{assignment}{This should be the outputted data frame from read_assignment.}

\item{sheaf}{This should be the sheaf structure from create_sheaf.}
}
\value{
This will return a sheaf assignment pair as an altered data frame.
}
\description{
This function will create a sheaf assignment pair.
This dictates a sheaf on a topological space paired with an assignment to the sheaf.
}
\details{
sheaf_assignment
}
\examples{
read_assignment(source = c("A", "B", "C"),
                values = c(2, 4, 3),
                variables = c("x", "y", "z")) -> assignment
X <- function(stalk){select("y")}
create_sheaf(restriction_maps = c(X), source = c("A"), destination = c("B")) -> sheaf
sheaf_assignment(assignment, sheaf)
}
